Update loginForms.css
Hey guys! I was feeling pretty encouraged from my last update so I spent a little time tonight playing with the login again. Some of you might have opinions against this, and that's fine. There's this debate out right now about "flat" design and how people don't know whether icons are buttons. It's much easier (more usable) for the user if it's obvious that something's a button. As it was, the 'Log In With Facebook' buttons were exactly the same style as the text input boxes. I made it more obvious that they're buttons by shading them.

I think the shading sets up a nice contrast, making the page more attractive. But if you guys want to be consistent with your Submit button or someone's a fan of the flat design fad, feel free to make adjustments :) Okay! Thanks everyone!