Update README.md
Improving the tone/voice of the opening paragraphs. Need one or two people to make sure that it still makes technical sense. 
I'm also trying to standardize the terms to lessen confusion and demonstrate integrity by referring to the other js systems by their preferred way.
Need a partner to help me with some tech phrasing and terms. 
The prerequisites are listed nicely, but I'd like to give a very short explanation why you need these functions. Even if they are a no-brainer for coders, a little explanation will tell your audience why it is important to the process of starting with MEAN.
I REALLY NEED a partner to help me with some of the other instructions. I think my questions are really simple. I just need to ask some questions to that I can make the instructions clearer.
Regards, Beth
Want to work on the "## Installation"  and "lines . Need a short chat with someone to